moved host to core in helios_sim
