+++ helios_sim/src/prelude.rs
@@ -7,7 +7,7 @@ pub use bevy::prelude::*;
 // pure types like `FrameHandle`, `Dynamics`, `Measurement`, etc.
 pub use helios_core::prelude::*;
 
-// Re-export common simulationecific types for easy access in other plugins.
+// Re-export common simulation-specific types for easy access in other plugins.
 pub use crate::cli::Cli;
 pub use crate::config::structs::{
     AgentConfig, RawScenarioConfig, ScenarioConfig, SensorConfig, Vehicle,
@@ -15,6 +15,7 @@ pub use crate::config::structs::{
 pub use crate::config::structs::{AtmosphereConfig, TerrainConfig};
 pub use crate::core::app_state::{AppState, SceneBuildSet, SimulationSet};
 pub use crate::core::components::{BoundingBox3D, SemanticLabel, TerrainMedium, WorldObjectType};
+pub use crate::core::host::{HeliosHost, Presentation};
 pub use crate::core::spawn_requests::SpawnAgentConfigRequest;
 pub use crate::core::transforms::{TfTree, TrackedFrame};
 
@@ -23,6 +24,3 @@ pub use crate::plugins::sensors::imu::ImuPlugin;
 pub use crate::plugins::vehicles::ackermann::{
     AckermannAdapterComponent, AckermannCarPlugin, AckermannOutputAdapter,
 };
-
-// host types
-pub use crate::host::{HeliosHost, Presentation};
